refactor: optimize layout and fix input warnings
